requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
cycler==0.10.0
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scipy==1.7.1
six==1.16.0
sqlparse==0.4.1
gunicorn==20.1.0
numpy==1.21.2
sqlparse==0.4.1
torch==1.9.0+cpu
torchvision==0.10.0+cpu
typing-extensions==3.10.0.0
whitenoise==5.3.0
